putting away work on valley_width_extraction for a while
The get wavelets algorithm currently is buggy for some centerlines and I'm not exactly sure why. Could try drawing centerlines manually. Also, each side of the valley is being sampled separately (instead of the entire perpendicular being used), which I think is causing errors in location.

Get_valleys_percentile probably has the best performance of all the methods., but when watersheds are too sinuous the algorithm misclassifies the side of the line that points fall on